refactor(anta): move anta.cli.cli module to anta.cli
